requirement.txt
elasticsearch==7.17.9
neo4j==5.14.1
pandas==2.1.4
python-docx==1.0.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
torch==2.1.2
transformers==4.36.2